Genrating oneOf with jaxrs-spec
