+++ modules/openapi-generator/src/main/resources/JavaJaxRS/spec/model.mustache
@@ -15,6 +15,10 @@ import javax.validation.Valid;
 {{#isEnum}}
 {{>enumOuterClass}}
 {{/isEnum}}
-{{^isEnum}}{{>pojo}}{{/isEnum}}
+{{^isEnum}}
+{{#vendorExtensions.x-is-one-of-interface}}{{>oneof_interface}}{{/vendorExtensions.x-is-one-of-interface}}{{^vendorExtensions.x-is-one-of-interface}}{{>pojo}}{{/vendorExtensions.x-is-one-of-interface}}
+{{/isEnum}}
+
 {{/model}}
 {{/models}}
+
